added new timed drive mode to swerve and auto
